+++ core/deployment/src/main/java/io/quarkus/deployment/Feature.java
@@ -104,6 +104,8 @@ public enum Feature {
     REST_CLIENT_MUTINY,
     REST_CLIENT_REACTIVE,
     REST_CLIENT_REACTIVE_JACKSON,
+    REST_CLIENT_REACTIVE_JAXB,
+    REST_CLIENT_REACTIVE_JSONB,
     REST_CLIENT_REACTIVE_KOTLIN_SERIALIZATION,
     SCALA,
     SCHEDULER,
