+++ extensions/resteasy-reactive/quarkus-resteasy-reactive-jaxb/deployment/src/main/java/io/quarkus/resteasy/reactive/jaxb/deployment/ResteasyReactiveJaxbProcessor.java
@@ -1,0 +1,14 @@
+package io.quarkus.resteasy.reactive.jaxb.deployment;
+
+import io.quarkus.deployment.Feature;
+import io.quarkus.deployment.annotations.BuildProducer;
+import io.quarkus.deployment.annotations.BuildStep;
+import io.quarkus.deployment.builditem.FeatureBuildItem;
+
+public class ResteasyReactiveJaxbProcessor {
+
+    @BuildStep
+    void feature(BuildProducer<FeatureBuildItem> feature) {
+        feature.produce(new FeatureBuildItem(Feature.RESTEASY_REACTIVE_JAXB));
+    }
+}
