+++ extensions/resteasy-reactive/quarkus-resteasy-reactive-jaxb/runtime/pom.xml
@@ -20,7 +20,7 @@
         </dependency>
         <dependency>
             <groupId>io.quarkus</groupId>
-            <artifactId>quarkus-jaxb</artifactId>
+            <artifactId>quarkus-resteasy-reactive-jaxb-common</artifactId>
         </dependency>
     </dependencies>
