--- extensions/resteasy-reactive/quarkus-resteasy-reactive-jaxb/runtime/src/main/java/io/quarkus/resteasy/reactive/jaxb/runtime/serialisers/JaxbMessageBodyWriter.java
@@ -1,67 +1,0 @@
-package io.quarkus.resteasy.reactive.jaxb.runtime.serialisers;
-
-import java.io.IOException;
-import java.io.OutputStream;
-import java.lang.annotation.Annotation;
-import java.lang.reflect.Type;
-import java.util.Map;
-
-import javax.ws.rs.WebApplicationException;
-import javax.ws.rs.core.HttpHeaders;
-import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.MultivaluedMap;
-import javax.xml.bind.JAXB;
-
-import org.jboss.resteasy.reactive.server.spi.ServerMessageBodyWriter;
-import org.jboss.resteasy.reactive.server.spi.ServerRequestContext;
-
-import io.vertx.core.MultiMap;
-
-public class JaxbMessageBodyWriter extends ServerMessageBodyWriter.AllWriteableMessageBodyWriter {
-
-    @Override
-    public void writeTo(Object o, Class<?> type, Type genericType, Annotation[] annotations, MediaType mediaType,
-            MultivaluedMap<String, Object> httpHeaders, OutputStream entityStream) throws WebApplicationException {
-        setContentTypeIfNecessary(httpHeaders);
-        JAXB.marshal(o, entityStream);
-    }
-
-    @Override
-    public void writeResponse(Object o, Type genericType, ServerRequestContext context)
-            throws WebApplicationException, IOException {
-        setContentTypeIfNecessary(context);
-        OutputStream stream = context.getOrCreateOutputStream();
-        JAXB.marshal(o, stream);
-        // we don't use try-with-resources because that results in writing to the http output without the exception mapping coming into play
-        stream.close();
-    }
-
-    private void setContentTypeIfNecessary(MultivaluedMap<String, Object> httpHeaders) {
-        Object contentType = httpHeaders.getFirst(HttpHeaders.CONTENT_TYPE);
-        if (isNotXml(contentType)) {
-            httpHeaders.putSingle(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_XML);
-        }
-    }
-
-    private void setContentTypeIfNecessary(ServerRequestContext context) {
-        String currentContentType = null;
-        Iterable<Map.Entry<String, String>> responseHeaders = context.serverResponse().getAllResponseHeaders();
-        if (responseHeaders instanceof MultiMap) {
-            currentContentType = ((MultiMap) responseHeaders).get(HttpHeaders.CONTENT_TYPE);
-        } else {
-            for (Map.Entry<String, String> entry : responseHeaders) {
-                if (entry.getKey().equalsIgnoreCase(HttpHeaders.CONTENT_TYPE)) {
-                    currentContentType = entry.getValue();
-                    break;
-                }
-            }
-        }
-        if (isNotXml(currentContentType)) {
-            context.serverResponse().setResponseHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_XML);
-        }
-    }
-
-    private boolean isNotXml(Object contentType) {
-        return contentType == null || !contentType.toString().contains("xml");
-    }
-}
