+++ extensions/resteasy-reactive/quarkus-resteasy-reactive-jsonb-common/runtime/src/main/resources/META-INF/quarkus-extension.yaml
@@ -1,0 +1,4 @@
+name: "RESTEasy Reactive JSON-B Common"
+artifact: ${project.groupId}:${project.artifactId}:${project.version}
+metadata:
+  unlisted: true
