+++ extensions/resteasy-reactive/quarkus-resteasy-reactive-jsonb/deployment/pom.xml
@@ -19,7 +19,7 @@
         </dependency>
         <dependency>
             <groupId>io.quarkus</groupId>
-            <artifactId>quarkus-jsonb-deployment</artifactId>
+            <artifactId>quarkus-resteasy-reactive-jsonb-common-deployment</artifactId>
         </dependency>
         <dependency>
             <groupId>io.quarkus</groupId>
