+++ extensions/resteasy-reactive/quarkus-resteasy-reactive-jsonb/deployment/src/main/java/io/quarkus/resteasy/reactive/jsonb/deployment/ResteasyReactiveJsonbProcessor.java
@@ -1,0 +1,20 @@
+package io.quarkus.resteasy.reactive.jsonb.deployment;
+
+import io.quarkus.deployment.Feature;
+import io.quarkus.deployment.annotations.BuildProducer;
+import io.quarkus.deployment.annotations.BuildStep;
+import io.quarkus.deployment.builditem.FeatureBuildItem;
+import io.quarkus.resteasy.reactive.common.deployment.ServerDefaultProducesHandlerBuildItem;
+
+public class ResteasyReactiveJsonbProcessor {
+
+    @BuildStep
+    void feature(BuildProducer<FeatureBuildItem> feature) {
+        feature.produce(new FeatureBuildItem(Feature.RESTEASY_REACTIVE_JSONB));
+    }
+
+    @BuildStep
+    ServerDefaultProducesHandlerBuildItem jsonDefault() {
+        return ServerDefaultProducesHandlerBuildItem.json();
+    }
+}
