+++ extensions/resteasy-reactive/quarkus-resteasy-reactive-jsonb/runtime/pom.xml
@@ -24,7 +24,7 @@
         </dependency>
         <dependency>
             <groupId>io.quarkus</groupId>
-            <artifactId>quarkus-jsonb</artifactId>
+            <artifactId>quarkus-resteasy-reactive-jsonb-common</artifactId>
         </dependency>
     </dependencies>
