+++ extensions/resteasy-reactive/rest-client-reactive-jaxb/runtime/src/main/resources/META-INF/quarkus-extension.yaml
@@ -1,0 +1,26 @@
+---
+artifact: ${project.groupId}:${project.artifactId}:${project.version}
+name: "REST Client Reactive JAXB"
+metadata:
+  keywords:
+  - "rest-client-jaxb"
+  - "rest-client"
+  - "web-client"
+  - "microprofile-rest-client"
+  - "jaxb"
+  - "xml"
+  - "resteasy-reactive"
+  categories:
+  - "web"
+  - "serialization"
+  status: "stable"
+  codestart:
+    name: "resteasy-reactive"
+    kind: "core"
+    languages:
+      - "java"
+      - "kotlin"
+      - "scala"
+    artifact: "io.quarkus:quarkus-project-core-extension-codestarts"
+  config:
+  - "quarkus.rest-client-reactive."
