+++ extensions/resteasy-reactive/rest-client-reactive-jsonb/runtime/src/main/resources/META-INF/quarkus-extension.yaml
@@ -1,0 +1,26 @@
+---
+artifact: ${project.groupId}:${project.artifactId}:${project.version}
+name: "REST Client Reactive JSON-B"
+metadata:
+  keywords:
+  - "rest-client-jsonb"
+  - "rest-client"
+  - "web-client"
+  - "microprofile-rest-client"
+  - "json"
+  - "jsonb"
+  - "resteasy-reactive"
+  categories:
+  - "web"
+  - "serialization"
+  status: "stable"
+  codestart:
+    name: "resteasy-reactive"
+    kind: "core"
+    languages:
+      - "java"
+      - "kotlin"
+      - "scala"
+    artifact: "io.quarkus:quarkus-project-core-extension-codestarts"
+  config:
+  - "quarkus.rest-client-reactive."
